[MOV] account_sale_timesheet,sale_project_account: remove useless bridges
Since the sale module depends on the account module, the bridges account_sale_timesheet
and sale_project_account can be merged with sale_project.

This commit removes those bridges and merges them with sale_project.

Task-2772395
