--- addons/account_sale_timesheet/i18n/account_sale_timesheet.pot
@@ -1,40 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_sale_timesheet
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~14.5\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-09-14 10:28+0000\n"
-"PO-Revision-Date: 2021-09-14 10:28+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_sale_timesheet
-#: model_terms:ir.ui.view,arch_db:account_sale_timesheet.project_project_view_form
-msgid ""
-"<span class=\"o_stat_text\">\n"
-"                            Invoices\n"
-"                        </span>"
-msgstr ""
-
-#. module: account_sale_timesheet
-#: model:ir.model.fields,field_description:account_sale_timesheet.field_project_project__invoice_count
-msgid "Invoice Count"
-msgstr ""
-
-#. module: account_sale_timesheet
-#: code:addons/account_sale_timesheet/models/project.py:0
-#, python-format
-msgid "Invoices"
-msgstr ""
-
-#. module: account_sale_timesheet
-#: model:ir.model,name:account_sale_timesheet.model_project_project
-msgid "Project"
-msgstr ""
