--- addons/account_sale_timesheet/views/project_project_views.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="project_project_view_form" model="ir.ui.view">
-        <field name="name">project.project.form.inherit</field>
-        <field name="model">project.project</field>
-        <field name="inherit_id" ref="project.edit_project"/>
-        <field name="priority">50</field>
-        <field name="arch" type="xml">
-            <xpath expr="//button[@name='action_view_analytic_account_entries']" position="after">
-                <button name="action_open_project_invoices" type="object" class="oe_stat_button" icon="fa-pencil-square-o"
-                        attrs="{'invisible': ['|', ('invoice_count', '=', 0), ('analytic_account_id', '=', False)]}"
-                        groups="account.group_account_readonly">
-                    <div class="o_field_widget o_stat_info">
-                        <span class="o_stat_value">
-                            <field name="invoice_count" nolabel="1"/>
-                        </span>
-                        <span class="o_stat_text">
-                            Invoices
-                        </span>
-                    </div>
-                </button>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
