--- addons/sale_project_account/i18n/sale_project_account.pot
@@ -1,41 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* sale_project_account
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~14.5\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-09-14 10:29+0000\n"
-"PO-Revision-Date: 2021-09-14 10:29+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: sale_project_account
-#: model_terms:ir.ui.view,arch_db:sale_project_account.project_project_view_form
-msgid ""
-"<span class=\"o_stat_text\">\n"
-"                            Vendor Bills\n"
-"                        </span>"
-msgstr ""
-
-#. module: sale_project_account
-#: model:ir.model,name:sale_project_account.model_project_project
-msgid "Project"
-msgstr ""
-
-#. module: sale_project_account
-#: model:ir.model.fields,field_description:sale_project_account.field_project_project__vendor_bill_count
-msgid "Vendor Bill Count"
-msgstr ""
-
-#. module: sale_project_account
-#: code:addons/sale_project_account/models/project.py:0
-#: code:addons/sale_project_account/models/project.py:0
-#, python-format
-msgid "Vendor Bills"
-msgstr ""
