--- addons/sale_project_account/views/project_project_views.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <record id="project_project_view_form" model="ir.ui.view">
-        <field name="name">project.project.form.inherit</field>
-        <field name="model">project.project</field>
-        <field name="inherit_id" ref="project.edit_project"/>
-        <field name="priority">10</field>
-        <field name="arch" type="xml">
-            <xpath expr="//button[@name='action_view_analytic_account_entries']" position="after">
-                <button name="action_open_project_vendor_bills" type="object" class="oe_stat_button" icon="fa-pencil-square-o" attrs="{'invisible': [('vendor_bill_count', '=', 0)]}" groups="account.group_account_readonly">
-                    <div class="o_field_widget o_stat_info">
-                        <span class="o_stat_value">
-                            <field name="vendor_bill_count" nolabel="1"/>
-                        </span>
-                        <span class="o_stat_text">
-                            Vendor Bills
-                        </span>
-                    </div>
-                </button>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
